+++ src/result/css/Product.module.css
@@ -1,5 +1,6 @@
 .container{
     width: 140px;
+    height:220px;
     position:relative;
     text-align: left;
     border-radius: 15px;
@@ -10,6 +11,7 @@
 }
 .img img{
     width: 140px;
+    height:140px;
     object-fit:cover;
     border-radius: 15px;
 }
@@ -35,7 +37,8 @@
 }
 .shopanddistance{
     display: flex;
-
+    margin:0px;
+    padding:0px;
 }
 .distance{
     border-left: solid;
@@ -51,6 +54,7 @@
 }
 .name{
     font-weight: bolder;
+    padding:0px;
 }
 .shopname{
     align-items: center;
