+++ src/result/css/Result.module.css
@@ -19,5 +19,27 @@
   display: flex;
   justify-content: center;
   flex-wrap: wrap;
-  margin-top: 3vh;
+  margin-top: 10px;
+}
+.sort{
+  border-top-style:solid;
+  border-bottom-style:solid;
+  border-width: 1px;
+  padding-top: 1vh;
+  padding-bottom: 1vh;
+  margin-top: 1vh;
+  width:100%;
+  display:flex;
+  justify-content: right;
+  align-items: center;
+}
+.sort img{
+  width:25px;
+}
+select{
+  border:none;
+  font-family: inherit;  
+  -webkit-appearance: none; 
+  -moz-appearance: none;
+  appearance: none;
 }
