--- src/result/css/Sort.module.css
@@ -1,24 +1,0 @@
-.container{
-    border-top-style:solid;
-    border-bottom-style:solid;
-    border-width: 1px;
-    padding-top: 1vh;
-    padding-bottom: 1vh;
-    margin-top: 1vh;
-    width:100%;
-    display:flex;
-    justify-content: right;
-    align-items: center;
-}
-.container img{
-    width:25px;
-}
-select{
-    width: 200px; 
-    padding: .8em .5em; 
-    border:none;
-    font-family: inherit;  
-    -webkit-appearance: none; 
-    -moz-appearance: none;
-    appearance: none;
-}
